git-ck: Fix un-committing with a "staged" commit but not an "unstaged" commit
The current behavior is such that if there is a "staged" commit (which
is assumed to be at element 1), but not an "unstaged" commit (which is
assumed to be at element 0), then the "unstaged" commit will not be
restored.

This has to do with assumptions about where each commit will be in the
produced list. This is a bad assumption, and would only get worse if
git-ck adds more commits.